chore: improve package structure
