+++ src/kaitai_awkward/_version.pyi
@@ -1,0 +1,4 @@
+from __future__ import annotations
+
+version: str
+version_tuple: tuple[int, int, int] | tuple[int, int, int, str, str]
